+++ OrganizeIt/app/src/main/java/com/example/organizeit/models/Drawer.kt
@@ -3,8 +3,8 @@ package com.example.organizeit.models
 import java.io.Serializable
 
 data class Drawer(
-    val id: Int? = null,
+    val id: Long? = null,
     var name: String,
     val order: Int,
-    val shelfId: Int? = null
+    val shelfId: Long? = null
 ) : Serializable
