+++ OrganizeIt/app/src/main/java/com/example/organizeit/models/Item.kt
@@ -1,11 +1,11 @@
 package com.example.organizeit.models
 
 data class Item(
-    val id: Int? = null,
+    val id: Long? = null,
     var name: String,
     var desc: String,
     var quantity: Double,
-    val drawerId: Int,
+    val drawerId: Long,
     var checkboxVisible: Boolean = false,
     var uncheck: Boolean = false
 )
